#41 Refactorization and files reorganization
