+++ application/src/server/services/NextBetQueryService.ts
