+++ application/src/server/services/NextBetService.ts
