+++ application/src/ui/icon/arrow-left.tsx
