+++ application/src/ui/icon/bug.tsx
