+++ application/src/ui/icon/check.tsx
