+++ application/src/ui/icon/close.tsx
