+++ application/src/ui/icon/delete.tsx
