+++ application/src/ui/icon/logout.tsx
