+++ application/src/ui/icon/mail.tsx
