+++ application/src/ui/icon/note.tsx
