+++ application/src/ui/icon/timer.tsx
