+++ application/src/ui/icon/verified.tsx
