+++ application/src/ui/icon/waving-hand.tsx
